Makefile.kdevops: add extra dependencies
To be able to run kdevops project, we need to make sure all dependencies
are met. For that, make sure the kdevops-deps includes min_deps target
(checks for sudo, make, nc and ansible-playbook) and menuconfig-deps
(installs ncurses).

Keep menuconfig-deps as target for users who do not want to run the full
kdevops dependency list check.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>