Fix remove_hosts() function
Commit 3b6fbaaaf1df ("update_ssh_config.py: add IP address on Host
line") added an IP address at the end of each Host line. This extra
field breaks the simple host entry matching logic in the
remove_hosts() function, making "--remove" a no-op.

If the ssh_config entries are not removed, the config file grows
without bounds. If new hostnames are added that have the same name
as existing entries, ssh to those hosts fails, which makes certain
Ansible tasks hang for a bit and then fail.

Fixes: 3b6fbaaaf1df ("update_ssh_config.py: add IP address on Host line")
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>