ansible_cfg: template boolean defaults must be "False"
I found that the "disable_skipped_hosts" Kconfig setting was not
working.

Using "output yaml" with a boolean behaves somewhat counter-
intuitively. When the user specifies "Yes, please set this", the
variable appears in .extra_vars_auto.yaml with a value of True.
When the user specifies "No, please do not set this" the variable
does not appear in .extra_vars_auto.yaml at all.

To make a template that uses these variables behave correctly, then,
the defaults/main.yml setting has to be False always. When the
Kconfig setting is enabled, that will override this default.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>