playbooks: Add a build_linux role
Currently, for cloud configurations or when 9p is disabled, kdevops
builds the test kernel on each test runner. This is inefficient,
and gets worse as the test matrix for a single kernel version scales
out.

Instead we want to build the test kernel once and make the build
artifacts available for test runners to install. This would be
similar to what KOTD does now, except it does not require setting up
a separate yum repo.

For this experiment, I've created a stripped down version of the
bootlinux role that has only the steps needed to build the kernel on
a target node. It adds some steps to create kernel packages and then
fetch them to the control host.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>