Kconfig: Define KDEVOPS_STORAGE_POOL_USER via Kconfig
Prepare to replace the "kdevops" string in the storage pool path
with the name of the user account running kdevops. First step is to
define a Kconfig variable for the storage pool user. For the moment,
the storage pool user name defaults to the same string it is
currently.

I would love for there to be a way to provide an actual pathname in
CONFIG_KDEVOPS_STORAGE_POOL_PATH. I haven't found a way so far.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>